requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.46.0
frozenlist==1.4.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
twilio==8.10.3
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
